Changed the direction of default_camera!
(1) The pose of the default camera is changed so it now sees Raven from the front (not the back)
    And the right and left arms are not flipped anymore!

(2) The Right Arm for raven used to have the tool shaft as a fixed joint.
    Changed it to continuous to match the left side.
    Now both arms can rotate their tool shafts ;)
    PS: this was an update from last commit, but I forgot to mention it :D Cheers!

What I've learned:
--------------------------------------------------------------------------
(1) The default_camera pose is set here:
    ambf > ambr_framework > afFramework.cpp (search for "default_camera)

(2) To set the camera pose, you need to set:
    camera position  (eye)
    look at position (target)
    the up vector    (orientation of camera)

(3) The camera is also movable through mouse commands:
    right button:  position
    middle scroll: zoom in and out
    left button:   orientation

(4) All the bodies are set initially wrt to world frame.
    So a serial link mechanism, it's way easier to change camera view rather than turn the serial link mechanism.
    Because changing the link_0 direction doesn't change the rest...
    It's scary.. don't try it!!!

(5) More info about the world frame.
    x: to the back of raven
    y: to the left of raven
    z: to the up direction of raven

What to explore:
----------------------------------------------------------------
(1) Allow multiple cameras?

(2) Can we move this existing default camera through commands?
    Or is it fixed... Let's figure it out ;)