Removed the need to comment out launch file for loading multi-bodies.
Previously to launch different multi-bodies, we were required to edit the launch.yaml file by un-commenting the required multibodies and commenting out the rest. This was rather cumbersome and unnecessary. With this update, one can use the -l command line option to pick which multi-bodies to. Multiple multibodies can be loaded by providing a comma separated list of indexes. E.g. ./ambf_simulator -l 1, 2, 6, 10 will load multibodies at indexes 1, 2, 6 and 10. If the -l flag is not used, the multibody at index 0 is loaded.