Segregated controllers for IIDs to allow coordinated shadow control
Previously the Input Interface Devices did not have a Cartesian controller
for the Simulated Dynamic End-Effectors (SDE). Instead, they relied on the
internal control of the SDE. With this commit, a new controller is defined for
the IID which is responsible for controlling the SDE. This allows multiple IIDs
to share a single SDE and take charge or specific Task space axes of freedom.