Can enable frame visualization for single body by using SHIFT + V
For this to work, first we have to select the body by enabling Mouse Picking Mode. This can be enabled by pressing 'P' on the keyboard once. Pressing 'P' again will toggle from Mouse Picking to Camera Pan. So select accordingly. Now you can click (and release) on an object to select it as a picked object. Finally, you can press SHIFT + N to toggle the drawing of frame of that body.