Using similarity transform for orientation offset
Previously, when an "orientation offset" was specified for the IID to pair an SDE, the orientation offset was imparted on the orientation of the SDE in world. This created confusion while setting the "orientation" component of the "location" as the SDE was not longer at the set "orientation" but at a mixture of the "location" "orientation" and the "orientation offset". Since the goal of the "orientation offset" is only to change the rotational mapping between the IID and SDE, this commit fixes the confusion.