First working implementation of preprocessing shaders
Description: These shaders would be run after the main rendering (to the screen) but before the rendering to the framebuffers. Thus the name preprocessing shaders. The current application is to generate segmentation masks for the depth point cloud. The depth pc uses the color image from the framebuffer to assign the color value to each point (from the corresponding color pixel).